cherry picking missing commits from sylabs sif repo
