+++ .github/workflows/ci.yml
@@ -1,5 +1,6 @@
 name: ci
 on:
+  workflow_dispatch: {}
   pull_request:
   push:
     branches:
@@ -22,7 +23,7 @@ jobs:
       - name: Set up Go
         uses: actions/setup-go@v1
         with:
-          go-version: '1.21.x'
+          go-version: '1.22.8'
 
       - name: Check go.mod and go.sum are tidy
         run: |
@@ -41,7 +42,7 @@ jobs:
       - name: Install Lint
         uses: golangci/golangci-lint-action@v2
         with:
-          version: v1.56
+          version: v1.61
           skip-pkg-cache: true
           skip-build-cache: true
