+++ pkg/integrity/digest_test.go
@@ -94,7 +94,6 @@ func TestNewLegacyDigest(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			d, err := newLegacyDigest(tt.ht, []byte(tt.text))
 			if got, want := err, tt.wantError; !errors.Is(got, want) {
@@ -160,7 +159,6 @@ func TestDigest_MarshalJSON(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			value, err := hex.DecodeString(tt.value)
 			if err != nil {
@@ -265,7 +263,6 @@ func TestDigest_UnmarshalJSON(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			var d digest
