+++ pkg/integrity/result.go
@@ -14,14 +14,13 @@ import (
 
 	"github.com/ProtonMail/go-crypto/openpgp"
 	"github.com/apptainer/sif/v2/pkg/sif"
-	"github.com/secure-systems-lab/go-securesystemslib/dsse"
 )
 
 // VerifyResult describes the results of an individual signature validation.
 type VerifyResult struct {
 	sig      sif.Descriptor
 	verified []sif.Descriptor
-	aks      []dsse.AcceptedKey
+	keys     []crypto.PublicKey
 	e        *openpgp.Entity
 	err      error
 }
@@ -38,11 +37,7 @@ func (r VerifyResult) Verified() []sif.Descriptor {
 
 // Keys returns the public key(s) used to verify the signature.
 func (r VerifyResult) Keys() []crypto.PublicKey {
-	keys := make([]crypto.PublicKey, 0, len(r.aks))
-	for _, ak := range r.aks {
-		keys = append(keys, ak.Public)
-	}
-	return keys
+	return r.keys
 }
 
 // Entity returns the signing entity, or nil if the signing entity could not be determined.
