+++ pkg/integrity/testdata/Test_dsseEncoder_signMessage/ED25519.golden
