+++ pkg/integrity/testdata/Test_dsseEncoder_signMessage/Multi.golden
