+++ pkg/integrity/testdata/Test_dsseEncoder_signMessage/RSA_SHA256.golden
