+++ pkg/integrity/verify.go
@@ -18,7 +18,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"sort"
+	"slices"
 	"strings"
 
 	"github.com/ProtonMail/go-crypto/openpgp"
@@ -576,9 +576,7 @@ func (v *Verifier) fingerprints(any bool) ([][]byte, error) {
 		}
 	}
 
-	sort.Slice(fps, func(i, j int) bool {
-		return bytes.Compare(fps[i], fps[j]) < 0
-	})
+	slices.SortFunc(fps, bytes.Compare)
 
 	return fps, nil
 }
