+++ pkg/sif/descriptor_input_test.go
@@ -199,8 +199,6 @@ func TestNewDescriptorInput(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
-		tt := tt
-
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
