+++ pkg/sif/descriptor_test.go
@@ -525,7 +525,6 @@ func TestDescriptor_GetIntegrityReader(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			d := Descriptor{
 				raw:        rd,
