+++ pkg/sif/load_test.go
@@ -37,8 +37,6 @@ func TestLoadContainerFromPath(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
-		tt := tt
-
 		t.Run(tt.name, func(t *testing.T) {
 			f, err := LoadContainerFromPath(tt.path, tt.opts...)
 			if err != nil {
@@ -70,8 +68,6 @@ func TestLoadContainer(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
-		tt := tt
-
 		t.Run(tt.name, func(t *testing.T) {
 			rw, err := os.Open(filepath.Join(corpus, "one-group.sif"))
 			if err != nil {
