+++ pkg/sif/sif_test.go
@@ -92,7 +92,6 @@ func TestHeader_GetIntegrityReader(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			b := bytes.Buffer{}
