+++ pkg/sif/testdata/TestDeleteObjects/DataTypeCompact.golden
